hybrid_pipeline/services/grayscale_service/requirements_aws.txt
boto3==1.28.85
botocore==1.31.85
python-dotenv
